gateway-2.0.20181213.1323

+++ cj.studio.gateway.console/src/cj/studio/gateway/socket/cable/IGatewaySocketCable.java
@@ -1,6 +1,6 @@
 package cj.studio.gateway.socket.cable;
 
-import cj.studio.ecm.graph.CircuitException;
+import cj.studio.ecm.net.CircuitException;
 import cj.ultimate.IClosable;
 import cj.ultimate.IDisposable;
 
@@ -63,7 +63,7 @@ public interface IGatewaySocketCable extends IClosable, IDisposable {
 
 	String protocol();
 
-	int getHeartbeat();
+	long getHeartbeat();
 
 	int workThreadCount();
 }
